Implement get_all_teams() to return all teams
I think this should work based on what we talked about in Slack, but it is untested so if you have a dev environment - it's worth running it through that first. Also, feel free to adjust the copy I added to the `get_teams()` method talking about `get_all_teams()`.